// ManagmentServer/API/UI/src/Components/WebHooks/WebHookList/__generated__/WebHookListDataFragment.graphql.ts
export type WebHookListDataFragment$data = {
  readonly webHooks: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"WebHookListItemDataFragment">;
      } | null;
    }> | null;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null;
    };
  } | null;
  readonly " $fragmentType": "WebHookListDataFragment";
};
export type WebHookListDataFragment$key = {
  readonly " $data"?: WebHookListDataFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"WebHookListDataFragment">;
};